Automatically adjust width of console output
The status line should be only one line. Previously it had a fixed
width of 99 characters. Now the width is set to the actual terminal
width. This value is determined again every time the status line is
updated, so that a running lending bot can be accessed from terminals
with different widths, like in a tmux setup.