Max to lent (#117)
Offers features for a lending volume conditional of the current rate. 

* Update lendingbot.py

Add the option to display the Account Estimated Earnings on BTC or in USD

* Update lendingbot.js

Add the option to display the Account Estimated Earnings on BTC or in USD

* Update README.md

Add the option to display the Account Estimated Earnings on BTC or in USD

* Update lendingbot.py

Changed to IP of the WebServer back to 127.0.0.1

* Update lendingbot.js

Change the decimal places of the USD to 8 again and add the @ at the USD price.

* Update lendingbot.py

Don't "require" an arg that has toggleable functionality.

* Update lendingbot.py

Remove the cancelLoanOffer test.

* Added the option to select the currency that the HTML Overview will present the earnings summary in.

* Max To Lend

* Max To Lend

* Max to lend fix

* Max to lend fix

* Max to lend fix

* Max To Lend

* Fix Do not make more than 6 api calls per second

If the coin have balance it tries to make the offer and make to many calls. 
Now if the activeBal is less than minLoanSize it update the totalCoins and skip to the next coin.

* Fix

if maxactive = 0 do not log the total coin. else log the total coin and then check if the activeBal is more than minLoanSize

* Fix

* Fix

* Fix

* begining with maxtolentrate

* begining with maxtolentrate

* test for maxtolentrate

* test for maxtolentrate

* test for maxtolentrate

* test for maxtolentrate

* fix for maxtolentrate

* fix for maxtolentrate

* fix for maxtolentrate

* fix for maxtolentrate

* fix for maxtolentrate

* fix for maxtolentrate

* fix for maxtolentrate

* fix for maxtolentrate

* fix for maxtolentrate

* fix for maxtolentrate Log

* max to lent rate

* Fix bugs

* change the log information