Modularise the bot code (#142)
* Module for webserver

* Create module subfolder

Clean up the main folder to look better and avoid confusion on what to click on. Tested and works.

* Moved some out of place code into Poloniex.py

A function that did nothing* other than call another function.

Also renamed it from poloniex.py to fit in with other modules naming styles.

* PEP8'd Poloniex.py

I like this change

* import Decimal instead of *

importing * is bad code

* Add Configuration.py

This compacted a lot of lines. Added ability for get() to have a default value.

* Fix the exception variable length issue

* MaxToLend module

* Change bot to api

Makes more sense

* Update configuration with maxtolend

* Fix saying "default" every other line

Happened in PyCharm's Python interpreter due to Consoleutils not being able to tell what OS it is on.

* Another change to config

Now we can set a config option as mandatory by setting the default to "None" so if it is commented out or removed when checked on then the bot will halt.

* Add Data module for just getting data

Functions that just return data or change log data (based on input or api)  go here

* line length fix

* Sane Sanity checks!

Now they are checked at the time of running and can be set inline with the Config.get() command

* Little more user friendly

Leave less to the imagination

* Wont lend under Poloniex min

Fixes #131

* Modularise Lending

Closes #138

* Missed some WebServer

* Simplify Logger construction

* Make lend_all() make more sense

Split it up and fixed some variables for logic

Also removed using an iterator instead of a for loop.

* Don't cancel stuck orders

Fixes #75

* Fix KeepStuckOrders option

oops

* Update docs

* keepstuckorders

* Max to lend code complexity (#143)

* Logger rename console to output.

* Reduce MaxToLend code complexity.

* Remove trailing semicolon

* Full lent -> lend