Overhaul lending module (#174)
* Overhaul lend_cur()

Made much of the creation of orders more direct.
TODO: Requires more work on when it reaches the end of how many offers it requested.
Fixes #165 - This error was due to splitting a satoshi in half, and they both wanted to round up. Unwanted result of fix is an occasional satoshi or two left in balance. TODO: Tack on extra sats to last offer.

* Implement increasing req limit

* Remove legacy function

Why is this still here

* Adjust to changing mins

Allows user to still pick any min size they want, but default is changed to 0.01
Will increase this min as required based off of API response.
Closes #180

* Handle outdated Numpy

So #184 won't happen again

* Optimize changes

* Fix finding remainder

* Handle short order book

* #189: Limit Poloniex API requests to 6 per second

Closes #189 by rnevet
Had to cherrypick this because the other branch included commits I was trying to force-remove.

* Bugfix

minloansize must be float

* Remove timestamps every request

* Fix chdir if example config not present

Fixes #192

* Make sure mindailyrate is the right var type

Should fix issues someone had in the comments of #191

* Add truncate() source

* Make sure everything is the right type 2

* Add config section for each coin and include minloansize for each coin (#193)

* Add config section for each coin (alt layout to coinconfig) and include minloansize for each coin

* Doc fixes, Configuration.py's get function and cast minloansize from cfg to Decimal

* Docs fix, add line to make 'warning' render correctly

* Look up conversion rates from blockchain.info for web server display (#194)

* Look up conversion rates from blockchain.info for web server display

* Use the API call instead of ticker, pass on non-existent currency

* Fix config issues mentioned in #91

Fix config parsing issues with logsize and webserverenabled

* Fix min_daily_rate check

* Don't fetch USDT from blockchain.info

* Proofreading fixes

Changed some undesirable behavior

* Fix maxactive

Mentioned in #195